bug修复

1.修改删除数据bug
2.修改入口的startTime，endTime的类型为long
3.修复cpu_idle查询接口

See merge request !10

+++ src/main/java/com/asiainfo/gim/monitor/task/GangliaDataCleanJob.java
@@ -9,9 +9,9 @@ public class GangliaDataCleanJob extends DaoCollection
 {
 
 	//历史数据保留的天数（以天为单位）
-	private int keepDuration;
+	private long keepDuration;
 	
-	public void setKeepDuration(int keepDuration)
+	public void setKeepDuration(long keepDuration)
 	{
 		this.keepDuration = keepDuration;
 	}
